+++ RELEASE-NOTES.md
@@ -1,1 +1,3 @@
 ## 0.26
+
+* #1 Support for collecting basic Amazon credentials